Add Support for Storing Photos
Patients should have a photo stored of them as well of size 512 x 512
pixels.
The photos are stored in the database as binary data (bytea) because:
- Want to maintain transactional integrity and ensure the image data is
  always consistent with database records.
- Have a moderate number of images and size is manageable.

To store a photo, firstly convert it to a base64 encoded string, and
store it in the photo field of the admin category in the json payload of
the Post request to the backend. The backend will then decode the base64
encoded string to binary format, and store it in the postgres database
in the binary format.

When getting a photo, the backend retrieves the photo as binary format,
then encodes it as a base64 string and sends it over in the json payload
response for the patient get request.
The photo field is not compulsory in the admin field, so there may be
patients that exist without a photo currently.

Additionally, the docker compose file is now replaced with a Dockerfile,
and the readme is updated accordingly. This is so that the same docker
docker container can be used for data access layer testing and
integration tests.