Add Login and Protected Routes
Backend now supports login, and paths are now protected. Authentication
is implemented via JWT (Json Web Tokens), which last for 8 hours.

Users will first have to log in by making a POST request, providing
the username and password in the json body (See the Postman request
for an example)
(Currently there is only 1 user with username and password of
"admin")

Upon successful login, the JSON response will return a token string,
that needs to be embedded in all subsequent requests to the backend via
the "Authorization" header.

The token expires after 8 hours, after which users will have to log in
again.

Currently, there is only 1 user available (admin), and it is not stored
in the database. Future implementations will either store the user in
the env file, or in the postgres database. Additionally, the password is
not stored in a hashed format.

Additionally, the secret key used to sign JWT claims is exposed in
middleware/auth.go. Future updates will require storing the key in an
env file to conceal it.