Add error messages for JSON Unmarshalling errors
Previously, only bindJSON errors could be detected by the backend, such
as in the case where entire fields were missing. JSON Unmarshalling
errors, such as when the type in a field did not match that of the Go
struct, return HTTP error 500 codes, which were not correct.

This fix detects and returns the appropriate JSON Unmarshalling errors
in the HTTP response body.