Refactor exportdb to use sql2csv library
Previously, the COPY method was invoked in the Docker container to
generate a csv file of the database at /tmp/output.csv, before accessing
it via a docker volume mount. This method is messy and complicated, so I
opted to use a simple sql2csv lib that gets the rows and writes to the
csv in the application code instead, removing the need for volumes.